backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
django-cors-headers==4.3.1